refactor(search): downsize UnifiedSearchParams.top_k to u16 + timeout_ms to u32 (TD-SEARCH-2 S3) (#1181)
Per-query thriftiness for the request struct (TD-SEARCH-2 S3 numeric slice):

- top_k: Option<usize> (16 B) -> Option<u16> (4 B). top_k is a small result
  count (realistic <= ~10^4, well under u16's 65_535 cap).
- timeout_ms: Option<u64> (16 B) -> Option<u32> (8 B). 4B ms (~49 days) is far
  beyond any query timeout.

UnifiedSearchParams is allocated/passed on every search, so this trims the
per-query struct. The small VALUE is preserved end-to-end; only the STORAGE
type shrinks.

Design: top_k is u16 in the request struct. Where it is read into engine
EXECUTION and used as a SIZE (BinaryHeap capacity, Vec::truncate, loop bound,
threshold), the local is usize and one `u16 as usize` cast sits at that
genuine size-boundary — usize is the correct Rust type for a size, and the cast
is transient stack (zero per-query cost). Set sites from usize sources cast
`as u16`. The ~100-file engine plumbing keeps top_k as usize (it is used as a
size there); u16 in the request struct is the per-query win.

core::search::SearchParams is a type alias for UnifiedSearchParams (one type).
Not on the OpenAPI spec (no ToSchema); serde ignores unknown fields, so the
wire is mixed-read-safe.

38 files: 2 field defs + Default in search_params.rs, the rest are boundary
casts at direct callers (set: `as u16`; size-read: `as usize`; timeout: `as u64`).

Verification (local, pre-CI):
- cargo check --all-targets (lib + tests + benches) OK
- cargo clippy --lib --bins (CI gate) OK
- cargo fmt --check OK
- cargo nextest -p proximadb-search-types: 40/40 pass (incl. serde round-trip)

Follow-up (separate): cascade u16 into other per-query structs (SearchContext,
SearchPlan, QueryContext) that still carry top_k: usize, to remove the residual
boundary casts.

+++ benches/bench_10_query_progressive.rs
@@ -95,7 +95,7 @@ async fn progressive_search_real(
     let search_params = Arc::new(SearchParams {
         vector: Some(query.to_vec()),
         query_vectors: None,
-        top_k: Some(k),
+        top_k: Some(k as u16),
         distance_metric: Some(DistanceMetric::Cosine),
         filter_expression: None,
         filters: None,
